add DataFrameStatSuite
