+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/ClientE2ETestSuite.scala
@@ -586,57 +586,6 @@ class ClientE2ETestSuite extends RemoteSparkSession {
       list.asScala.map(kv => Row(kv.key, kv.value)),
       session.createDataFrame(list.asScala.toSeq))
   }
-
-  test("approximate quantile") {
-    val session = spark
-    import session.implicits._
-
-    val n = 1000
-    val df = Seq.tabulate(n + 1)(i => (i, 2.0 * i)).toDF("singles", "doubles")
-
-    val q1 = 0.5
-    val q2 = 0.8
-    val epsilons = List(0.1, 0.05, 0.001)
-
-    for (epsilon <- epsilons) {
-      val Array(single1) = df.stat.approxQuantile("singles", Array(q1), epsilon)
-      val Array(double2) = df.stat.approxQuantile("doubles", Array(q2), epsilon)
-      // Also make sure there is no regression by computing multiple quantiles at once.
-      val Array(d1, d2) = df.stat.approxQuantile("doubles", Array(q1, q2), epsilon)
-      val Array(s1, s2) = df.stat.approxQuantile("singles", Array(q1, q2), epsilon)
-
-      val errorSingle = 1000 * epsilon
-      val errorDouble = 2.0 * errorSingle
-
-      assert(math.abs(single1 - q1 * n) <= errorSingle)
-      assert(math.abs(double2 - 2 * q2 * n) <= errorDouble)
-      assert(math.abs(s1 - q1 * n) <= errorSingle)
-      assert(math.abs(s2 - q2 * n) <= errorSingle)
-      assert(math.abs(d1 - 2 * q1 * n) <= errorDouble)
-      assert(math.abs(d2 - 2 * q2 * n) <= errorDouble)
-
-      // Multiple columns
-      val Array(Array(ms1, ms2), Array(md1, md2)) =
-        df.stat.approxQuantile(Array("singles", "doubles"), Array(q1, q2), epsilon)
-
-      assert(math.abs(ms1 - q1 * n) <= errorSingle)
-      assert(math.abs(ms2 - q2 * n) <= errorSingle)
-      assert(math.abs(md1 - 2 * q1 * n) <= errorDouble)
-      assert(math.abs(md2 - 2 * q2 * n) <= errorDouble)
-    }
-
-    // quantile should be in the range [0.0, 1.0]
-    val e = intercept[IllegalArgumentException] {
-      df.stat.approxQuantile(Array("singles", "doubles"), Array(q1, q2, -0.1), epsilons.head)
-    }
-    assert(e.getMessage.contains("percentile should be in the range [0.0, 1.0]"))
-
-    // relativeError should be non-negative
-    val e2 = intercept[IllegalArgumentException] {
-      df.stat.approxQuantile(Array("singles", "doubles"), Array(q1, q2), -1.0)
-    }
-    assert(e2.getMessage.contains("Relative Error must be non-negative"))
-  }
 }
 
 private[sql] case class MyType(id: Long, a: Double, b: Double)
